requirements.txt
hydra-core>=1.3.2
torch>=2.5.0
transformers[torch]>=4.49.0
einops>=0.8.1
pyarrow>=19.0.1
psycopg2>=2.9.10
datasets>=3.3.2
hydra-trainer>=0.1.2
optuna
wandb
ninja
pytest
pandarallel>=1.6.5
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
ipykernel
